+++ examples/38_cookies/waku.config.ts
@@ -1,9 +1,10 @@
-/** @type {import('waku/config').Config} */
-export default {
+import { defineConfig } from 'waku/config';
+
+export default defineConfig({
   middleware: () => [
     import('waku/middleware/context'),
     import('./src/middleware/cookie.js'),
     import('waku/middleware/dev-server'),
     import('waku/middleware/handler'),
   ],
-};
+});
